+++ examples/std/qbit.go
@@ -10,15 +10,19 @@ import (
 )
 
 func QBit() error {
-	conn, err := GetStdOpenDBConnection(clickhouse.Native, clickhouse.Settings{
-		// QBit is an experimental feature in ClickHouse
-		"enable_qbit_type": 1,
-	}, nil, nil)
+	conn, err := GetStdOpenDBConnection(clickhouse.Native, nil, nil, nil)
 	if err != nil {
 		return err
 	}
+	if !CheckMinServerVersion(conn, 25, 10, 0) {
+		fmt.Print("unsupported clickhouse version for QBit type")
+		return nil
+	}
 	ctx := context.Background()
-
+	ctx = clickhouse.Context(ctx, clickhouse.WithSettings(clickhouse.Settings{
+		// QBit is an experimental feature in ClickHouse
+		"allow_experimental_qbit_type": 1,
+	}))
 	_, err = conn.ExecContext(ctx, "DROP TABLE IF EXISTS example")
 	if err != nil {
 		return err
@@ -106,7 +110,7 @@ func QBit() error {
 	var searchQuery = `
 		SELECT
 			id,
-			L2DistanceTransposed(embedding, ?::Array(Float32)) as distance
+			L2DistanceTransposed(embedding, ?::Array(Float32), 32) as distance
 		FROM example
 		ORDER BY distance ASC
 		LIMIT 3
